requirements.txt
wheel
Jinja2==3.0.3
jira==3.1.1
PyYAML==6.0
yamale==4.0.3
rich==12.0.0
